refrigerator_label_pi/requirements.txt
brother-ql==0.9.4
Pillow==8.1.2
qrcode==7.3.1
requests==2.25.1
python-dotenv